Use the correct frame number to extract commands from
This fixes using the incorrect frame when instantiating CLI without
passing a commands parameter. Prior to this fix, the module where the
CLI class is defined was being used, when what needs to be used is the
one where the instantiation is happening.

With this fix, the following idiom now works as expected:

      # This is in a separate module
      from cmdcmd import CLI, Command

      class cmd_foo(Command):
          def run(self):
              pass

      if __name__ == "__main__":
          CLI().run()