Avoid builtin shadowing Flake8 warnings
There two different things here: In the signature of the Option
constructor, we do actually want the names to be "help" and "type",
even when they are indeed shadowing builtins. In the rest of the code
we do not need to have those, so the variable is renamed to "_help".